+++ tee-worker/litentry/core/credentials/src/nodereal/crypto_summary/mod.rs
@@ -15,88 +15,376 @@
 // along with Litentry.  If not, see <https://www.gnu.org/licenses/>.
 
 use crate::*;
+use lc_data_providers::{
+	achainable::web3_network_to_chain,
+	nodereal_jsonrpc::{
+		FungibleApiList, GetNFTHoldingsParam, NftApiList, NoderealChain, NoderealJsonrpcClient,
+	},
+};
+use litentry_primitives::{ErrorDetail, IntoErrorDetail};
+use serde::{Deserialize, Serialize};
+use std::{string::String, vec, vec::Vec};
 
-const VC_CRYPTO_SUMMARY_DESCRIPTIONS: &str = "DESCRIPTION placeholder";
-const VC_CRYPTO_SUMMARY_TYPE: &str = "TYPE placeholder";
+use self::summary::{CRYPTO_SUMMARY_NFT_ADDRESSES_ETH, CRYPTO_SUMMARY_TOKEN_ADDRESSES_ETH};
 
-pub const CRYPTO_SUMMARY_TOKEN_ADDRESSES_ETH: [(&str, &str); 15] = [
-	("0x6982508145454Ce325dDbE47a25d4ec3d2311933", "PEPE"),
-	("0x95aD61b0a150d79219dCF64E1E6Cc01f0B64C4cE", "SHIB"),
-	("0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984", "UNI"),
-	("0x0000000000000000000000000000000000000000", "ETH"),
-	("0xB8c77482e45F1F44dE1745F52C74426C631bDD52", "BNB"),
-	("0x514910771AF9Ca656af840dff83E8264EcF986CA", "LINK"),
-	("0x5283d291dbcf85356a21ba090e6db59121208b44", "BLUR"),
-	("0xB50721BCf8d664c30412Cfbc6cf7a15145234ad1", "ARB"),
-	("0x0d8775f648430679a709e98d2b0cb6250d2887ef", "BAT"),
-	("0xe28b3B32B6c345A34Ff64674606124Dd5Aceca30", "INJ"),
-	("0x7fc66500c84a76ad7e9c93437bfc5ac33e2ddae9", "AAVE"),
-	("0x163f8c2467924be0ae7b5347228cabf260318753", "WLD"),
-	("0x50D1c9771902476076eCFc8B2A83Ad6b9355a4c9", "FTT"),
-	("0x152649eA73beAb28c5b49B26eb48f7EAD6d4c898", "CAKE"),
-	("0xb59490aB09A0f526Cc7305822aC65f2Ab12f9723", "LIT"),
-];
+pub mod summary;
 
-pub const CRYPTO_SUMMARY_NFT_ADDRESSES_ETH: [(&str, &str); 15] = [
-	("0x9401518f4EBBA857BAA879D9f76E1Cc8b31ed197", "The Weirdo Ghost Gang"),
-	("0x23581767a106ae21c074b2276D25e5C3e136a68b", "Moonbirds"),
-	("0x142e03367eDE17Cd851477A4287D1F35676E6dC2", "Yogapetz"),
-	("0x59325733eb952a92e069C87F0A6168b29E80627f", "Mocaverse"),
-	("0xeC19CAeF9C046f5f87A497154766742ab9C90820", "y00ts"),
-	("0x5CC5B05a8A13E3fBDB0BB9FcCd98D38e50F90c38", "The Sandbox"),
-	("0x596a5CD859AD53fEc23Cd3fCD77522f0B407920d", "MATR1X KUKU"),
-	("0x8a90CAb2b38dba80c64b7734e58Ee1dB38B8992e", "Doodles"),
-	("0x49cF6f5d44E70224e2E23fDcdd2C053F30aDA28B", "CloneX"),
-	("0xd774557b647330C91Bf44cfEAB205095f7E6c367", "Nakamigos"),
-	("0x6339e5E072086621540D0362C4e3Cea0d643E114", "Opepen"),
-	("0xe1dC516B1486Aba548eecD2947A11273518434a4", "The Grapes"),
-	("0x769272677faB02575E84945F03Eca517ACc544C", "The Captainz"),
-	("0x39ee2c7b3cb80254225884ca001F57118C8f21B6", "The Potatoz"),
-	("0x7Bd29408f11D2bFC23c34f18275bBf23bB716Bc7", "Meebits"),
-];
-
-pub trait CryptoSummaryCredentialUpdate {
-	fn update_crypto_summary_credential(&mut self, is_empty: bool);
+/*
+SUMMARY: {
+	TOKEN: [
+		{
+			network: BSC,
+			list: [
+				{
+					name: PEPE,
+					address: 0x123,
+				},
+				{
+					name: SHIB,
+					address: 0x95aD61b0a150d79219dCF64E1E6Cc01f0B64C4cE ,
+				},
+				//...
+			]
+		},
+		{
+			network: Ethereum,
+			list: [
+				{
+					name: PEPE,
+					address: 0x123,
+				},
+				{
+					name: SHIB,
+					address: 0x95aD61b0a150d79219dCF64E1E6Cc01f0B64C4cE ,
+				},
+				//...
+			]
+		}
+	],
+	NFT: [
+		{
+			network: Ethereum,
+			list: [
+				{
+					name: Moonbirds,
+					address: 0x23581767a106ae21c074b2276D25e5C3e136a68b
+				}
+			]
+		}
+	]
 }
 
-impl CryptoSummaryCredentialUpdate for Credential {
-	fn update_crypto_summary_credential(&mut self, _is_empty: bool) {
-		// let found_tokens: Vec<String> = items
-		// 	.iter()
-		// 	.filter_map(|item| {
-		// 		let lowercase_token_address = item.token_address.to_lowercase(),
-		// 		CRYPTO_SUMMARY_TOKEN_ADDRESSES
-		// 			.get(&lowercase_token_address)
-		// 			.map(|name| name.to_string())
-		// 	})
-		// 	.collect(),
+*/
 
-		// 	let mut and_logic = AssertionLogic::new_and(),
-		// 	AssertionLogic::new_item("$network", Op::Equal, &ETH),
-		// 	AssertionLogic::new_item("$network", Op::Equal, &Polkadot),
-		// 	AssertionLogic::new_item("$network", Op::Equal, &Polygon),
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ResponseToken {
+	token_address: String,
+	token_balance: String,
+	token_name: String,
+	token_symbol: String,
+}
 
-		// 	AssertionLogic::new_item("$token", Op::Equal, ""),
+#[derive(Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ResponseTokenResult {
+	pub total_count: String,
+	pub native_token_balance: String,
+	pub details: Vec<ResponseToken>,
+}
 
-		// // from_date's Op is ALWAYS Op::LessThan
-		// let from_date_logic = AssertionLogic::new_item("$from_date", Op::LessThan, from_date),
+#[derive(Serialize, Deserialize, Debug, Default)]
+pub struct NameAndAddress {
+	pub name: String,
+	pub address: String,
+}
 
-		// // minimum_amount' Op is ALWAYS Op::Equal
-		// let minimum_amount_logic =
-		// 	AssertionLogic::new_item("$minimum_amount", Op::Equal, minimum_amount),
+impl NameAndAddress {
+	pub fn new(name: &str, address: &str) -> Self {
+		Self { name: name.to_string(), address: address.to_string() }
+	}
+}
 
-		// // to_date's Op is ALWAYS Op::GreaterEq
-		// let to_date = format_assertion_to_date(),
-		// let to_date_logic = AssertionLogic::new_item("$to_date", Op::GreaterEq, &to_date),
+#[derive(Serialize, Deserialize, Debug, Default)]
+pub struct Item {
+	pub network: String,
+	pub list: Vec<NameAndAddress>,
+}
 
-		// let assertion = AssertionLogic::new_and()
-		// 	.add_item(minimum_amount_logic)
-		// 	.add_item(from_date_logic)
-		// 	.add_item(to_date_logic),
+impl Item {
+	pub fn new(network: String, list: Vec<NameAndAddress>) -> Self {
+		let mut item = Item::default();
 
-		// self.credential_subject.assertions.push(assertion),
-		// self.credential_subject.values.push(value),
+		if !list.is_empty() {
+			item.network = network;
+			item.list = list;
+		}
 
-		self.add_subject_info(VC_CRYPTO_SUMMARY_DESCRIPTIONS, VC_CRYPTO_SUMMARY_TYPE);
+		item
+	}
+
+	pub fn is_empty(&self) -> bool {
+		self.list.is_empty()
+	}
+}
+
+#[derive(Serialize, Deserialize, Debug, Default)]
+pub struct TokenAndNFT {
+	#[serde(rename = "TOKEN")]
+	#[serde(skip_serializing_if = "Vec::is_empty")]
+	pub token: Vec<Item>,
+
+	#[serde(rename = "NFT")]
+	#[serde(skip_serializing_if = "Vec::is_empty")]
+	pub nft: Vec<Item>,
+}
+
+impl TokenAndNFT {
+	pub fn add_token(&mut self, item: Item) {
+		if !item.is_empty() {
+			self.token.push(item);
+		}
+	}
+
+	pub fn add_nft(&mut self, item: Item) {
+		if !item.is_empty() {
+			self.nft.push(item);
+		}
+	}
+}
+
+#[derive(Serialize, Deserialize, Debug, Default)]
+pub struct SummaryHoldings {
+	#[serde(rename = "SUMMARRY")]
+	pub summary: TokenAndNFT,
+}
+
+impl SummaryHoldings {
+	pub fn is_empty(&self) -> bool {
+		self.summary.token.is_empty() && self.summary.nft.is_empty()
+	}
+
+	pub fn construct(bsc_token: &[bool], eth_token: &[bool], eth_nft: &[bool]) -> Self {
+		let bsc_token_list = Self::collect_list(&CRYPTO_SUMMARY_TOKEN_ADDRESSES_ETH, bsc_token);
+		let eth_token_list = Self::collect_list(&CRYPTO_SUMMARY_TOKEN_ADDRESSES_ETH, eth_token);
+		let eth_nft_list = Self::collect_list(&CRYPTO_SUMMARY_NFT_ADDRESSES_ETH, eth_nft);
+
+		let mut token_and_nft = TokenAndNFT::default();
+		token_and_nft
+			.add_token(Item::new(web3_network_to_chain(&Web3Network::Bsc), bsc_token_list));
+		token_and_nft
+			.add_token(Item::new(web3_network_to_chain(&Web3Network::Ethereum), eth_token_list));
+		token_and_nft
+			.add_nft(Item::new(web3_network_to_chain(&Web3Network::Ethereum), eth_nft_list));
+
+		SummaryHoldings { summary: token_and_nft }
+	}
+
+	fn collect_list(source: &[(&str, &str); 15], flags: &[bool]) -> Vec<NameAndAddress> {
+		let mut list = vec![];
+		for (index, is_holder) in flags.iter().enumerate() {
+			if *is_holder {
+				let (address, name) = source[index];
+				let name_and_address = NameAndAddress::new(name, address);
+				list.push(name_and_address)
+			}
+		}
+
+		list
+	}
+}
+
+pub struct CryptoSummaryClient {
+	pub eth_client: NoderealJsonrpcClient,
+	pub bsc_client: NoderealJsonrpcClient,
+}
+
+impl Default for CryptoSummaryClient {
+	fn default() -> Self {
+		Self::new()
+	}
+}
+
+impl CryptoSummaryClient {
+	pub fn new() -> Self {
+		let eth_client = NoderealJsonrpcClient::new(NoderealChain::Eth);
+		let bsc_client = NoderealJsonrpcClient::new(NoderealChain::Bsc);
+
+		Self { eth_client, bsc_client }
+	}
+
+	pub fn logic(
+		&mut self,
+		identities: &Vec<(Web3Network, Vec<String>)>,
+	) -> core::result::Result<SummaryHoldings, ErrorDetail> {
+		// Corresponds one-to-one with CRYPTO_SUMMARY_TOKEN_ADDRESSES_ETH
+		let mut flag_bsc_token: [bool; 15] = [false; 15];
+		let mut flag_eth_token: [bool; 15] = [false; 15];
+		let mut flag_eth_nft: [bool; 15] = [false; 15];
+
+		for (network, addresses) in identities {
+			if *network == Web3Network::Bsc {
+				// Token
+				// If these BSC Tokens are all included, there is no need to continue searching
+				if flag_bsc_token.contains(&false) {
+					for address in addresses {
+						let res = self
+							.bsc_client
+							.get_token_holdings(address)
+							.map_err(|e| e.into_error_detail())?;
+
+						let result: ResponseTokenResult = serde_json::from_value(res.result)
+							.map_err(|_| ErrorDetail::ParseError)?;
+						let mut token_addresses = vec![];
+						result.details.iter().for_each(|detail| {
+							token_addresses.push(detail.token_address.clone());
+						});
+						Self::update_holding_flag(
+							&mut flag_bsc_token,
+							CRYPTO_SUMMARY_TOKEN_ADDRESSES_ETH,
+							&token_addresses,
+						);
+					}
+				}
+			}
+
+			if *network == Web3Network::Ethereum {
+				for address in addresses {
+					// Tokens
+					// If these ETH ERC20 are all included, there is no need to continue searching
+					if flag_eth_token.contains(&false) {
+						let res_token = self
+							.eth_client
+							.get_token_holdings(address)
+							.map_err(|e| e.into_error_detail())?;
+						let result: ResponseTokenResult = serde_json::from_value(res_token.result)
+							.map_err(|_| ErrorDetail::ParseError)?;
+
+						let mut token_addresses = vec![];
+						result.details.iter().for_each(|detail| {
+							token_addresses.push(detail.token_address.clone());
+						});
+						Self::update_holding_flag(
+							&mut flag_eth_token,
+							CRYPTO_SUMMARY_TOKEN_ADDRESSES_ETH,
+							&token_addresses,
+						);
+					}
+
+					// NFT
+					// If these NFTs are all included, there is no need to continue searching
+					if flag_eth_nft.contains(&false) {
+						let param = GetNFTHoldingsParam {
+							account_address: address.to_string(),
+							token_type: "ERC721".to_string(),
+							page: 1,
+							page_size: 100,
+						};
+
+						let res_nft = self
+							.eth_client
+							.get_nft_holdings(&param)
+							.map_err(|e| e.into_error_detail())?;
+						let details = res_nft.details;
+
+						let mut nft_addresses = vec![];
+						details.iter().for_each(|detail| {
+							nft_addresses.push(detail.token_address.clone());
+						});
+
+						Self::update_holding_flag(
+							&mut flag_eth_nft,
+							CRYPTO_SUMMARY_NFT_ADDRESSES_ETH,
+							&nft_addresses,
+						)
+					}
+				}
+			}
+		}
+
+		Ok(SummaryHoldings::construct(&flag_bsc_token, &flag_eth_token, &flag_eth_nft))
+	}
+
+	fn update_holding_flag(
+		flag_array: &mut [bool],
+		source: [(&'static str, &'static str); 15],
+		token_addresses: &[String],
+	) {
+		for (index, is_holder) in flag_array.iter_mut().enumerate() {
+			if !*is_holder {
+				let (token_address, _token_name) = source[index];
+				if token_addresses.contains(&token_address.to_string()) {
+					*is_holder = true;
+				}
+			}
+		}
+	}
+}
+
+#[cfg(test)]
+mod tests {
+	use crate::nodereal::crypto_summary::summary::CRYPTO_SUMMARY_NFT_ADDRESSES_ETH;
+	use litentry_primitives::Web3Network;
+
+	use super::{CryptoSummaryClient, SummaryHoldings};
+
+	#[test]
+	fn summary_construct_works() {
+		let flag_bsc_token = [
+			false, true, true, true, true, true, true, true, true, true, true, true, true, true,
+			true,
+		];
+		let flag_eth_token = [
+			false, false, false, false, false, false, false, false, false, false, false, false,
+			false, false, false,
+		];
+		let flag_eth_nft = [
+			true, false, false, false, false, false, false, false, false, false, false, false,
+			false, false, false,
+		];
+
+		let summary = SummaryHoldings::construct(&flag_bsc_token, &flag_eth_token, &flag_eth_nft);
+		assert!(!summary.is_empty());
+	}
+
+	#[test]
+	fn update_holding_flag_works() {
+		let nft_addresses = vec![
+			"0x9401518f4EBBA857BAA879D9f76E1Cc8b31ed197".to_string(),
+			"0x6339e5E072086621540D0362C4e3Cea0d643E114".to_string(),
+		];
+		let mut flag_eth_nft = [
+			false, false, false, false, false, false, false, false, false, false, false, false,
+			false, false, false,
+		];
+		CryptoSummaryClient::update_holding_flag(
+			&mut flag_eth_nft,
+			CRYPTO_SUMMARY_NFT_ADDRESSES_ETH,
+			&nft_addresses,
+		);
+		assert!(flag_eth_nft.contains(&true));
+	}
+
+	#[test]
+	fn logic_works() {
+		// Ethereum:
+		// 0x9cdF4E1347328416daF17335CaF6A314201CC1Dd -> 32000000
+		// 0xF977814e90dA44bFA03b6295A0616a897441aceC -> 31000000
+
+		// BSC:
+		// 0x11fBffc1F3dF23E7219e2B3036fe2A12C10cD3AD -> 2200000
+		let identities = vec![
+			(
+				Web3Network::Ethereum,
+				vec![
+					"0x9cdF4E1347328416daF17335CaF6A314201CC1Dd".to_string(),
+					"0xF977814e90dA44bFA03b6295A0616a897441aceC".to_string(),
+				],
+			),
+			(Web3Network::Bsc, vec!["0x11fBffc1F3dF23E7219e2B3036fe2A12C10cD3AD".to_string()]),
+		];
+
+		let summary = CryptoSummaryClient::new().logic(&identities).unwrap();
+		println!("summary: {:?}", summary);
 	}
 }
